refactor: export and use core-snapshots types
