+++ packages/core-snapshots-cli/package.json
@@ -44,6 +44,7 @@
     "dependencies": {
         "@arkecosystem/core-interfaces": "^2.1.0",
         "@arkecosystem/core-container": "^2.1.0",
+        "@arkecosystem/core-snapshots": "^2.1.0",
         "@types/boom": "^7.2.1",
         "@types/cli-progress": "^1.8.0",
         "@types/commander": "^2.12.2",
