+++ packages/core-snapshots-cli/src/commands/truncate.ts
@@ -1,6 +1,7 @@
 import { app } from "@arkecosystem/core-container";
+import { SnapshotManager } from "@arkecosystem/core-snapshots";
 
 export async function truncateSnapshot(options) {
-    const snapshotManager = app.resolvePlugin("snapshots");
+    const snapshotManager = app.resolvePlugin<SnapshotManager>("snapshots");
     await snapshotManager.truncateChain();
 }
