+++ packages/core-snapshots/package.json
@@ -6,7 +6,8 @@
         "Kristjan Košič <chris@ark.io>"
     ],
     "license": "MIT",
-    "main": "dist/index.js",
+    "main": "dist/index",
+    "types": "dist/index",
     "files": [
         "dist"
     ],
